Task 03 [03]
Add advanced checking for file path
