+++ 03-streams/03-file-server-get/isPathOfFirstLevel.js
@@ -9,6 +9,10 @@
  * @return {boolean}
  */
 function isPathOfFirstLevel(path) {
+  if (!path) {
+    return false;
+  }
+
   return !/[\/\\]/.test(path);
 }
